% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacie.R
\name{find_files}
\alias{find_files}
\title{Find Files
 Finding the files in the directory which match the filename}
\usage{
find_files(filename, path)
}
\arguments{
\item{filename}{The name of the file you are searching for in the path}

\item{directory}{The path you are searching for the file in}
}
\value{
gives a vector of strings with the filenames in
}
\description{
Find Files
 Finding the files in the directory which match the filename
}
