% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacie.R
\name{get_face}
\alias{get_face}
\title{Get Face Function
  gets the full  path of the face of the glacier, the most recent snapshot of the data}
\usage{
get_face(filename, path = getwd(), legacy_time = 10)
}
\arguments{
\item{legacy_time}{}
}
\description{
Get Face Function
  gets the full  path of the face of the glacier, the most recent snapshot of the data
}
