% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacie.R
\name{get_max}
\alias{get_max}
\title{Get Max
  Get the most recent date of the date stamps used in the files}
\usage{
get_max(files)
}
\arguments{
\item{files}{}
}
\description{
Get Max
  Get the most recent date of the date stamps used in the files
}
