% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_and_surface.R
\name{iceBuildGlacierPath}
\alias{iceBuildGlacierPath}
\title{Path builder for glacier}
\usage{
iceBuildGlacierPath(path, file, extension)
}
\arguments{
\item{extension}{}
}
\description{
Path builder for glacier
}
