% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_and_surface.R
\name{iceGetCleanPath}
\alias{iceGetCleanPath}
\title{Getter function to find the clean path to the configured data location}
\usage{
iceGetCleanPath()
}
\description{
Getter function to find the clean path to the configured data location
}
