% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_and_surface.R
\name{iceSurface}
\alias{iceSurface}
\title{Surface}
\usage{
iceSurface(object, show = FALSE)
}
\description{
Function search the processed data directory in the config file for the named file and pull it up
}
