% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacie.R
\name{latest_file}
\alias{latest_file}
\title{Latest File
  Function to get the filename of the most recent file}
\usage{
latest_file(files)
}
\arguments{
\item{files}{A set of filenames}
}
\value{
The full filename of the most recent file in that set
}
\description{
Latest File
  Function to get the filename of the most recent file
}
