% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacie.R
\name{refresh_sheet}
\alias{refresh_sheet}
\title{Refresh Sheet Function
  Gets the most recent format of an object, if that object follows glacier
  naming conventions}
\usage{
refresh_sheet(varname, path = getwd(), flat_type = ".csv")
}
\arguments{
\item{flat_type}{}
}
\description{
Refresh Sheet Function
  Gets the most recent format of an object, if that object follows glacier
  naming conventions
}
