% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacie.R
\name{find_files}
\alias{find_files}
\title{Find Files
 Finding the files in the directory which match the filename}
\usage{
find_files(filename, path)
}
\arguments{
\item{directory}{}
}
\description{
Find Files
 Finding the files in the directory which match the filename
}
